refactor(operator_commands_dashboard): extract tmux WebSocket + listing handlers
Move TmuxSession struct, parse_tmux_sessions, sanitize_tmux_ident,
run_tmux_list_for_host, azlin_tmux_sessions, ws_tmux_attach_handler,
and handle_tmux_attach_ws from routes.rs into a new sibling tmux.rs
module.

- routes.rs: 3773 -> 3383
- tmux.rs (NEW, ~405)

Refs #1266
