+++ src/operator_commands_dashboard/tests_routes_a.rs
@@ -1,7 +1,7 @@
 #[cfg(test)]
 mod tests {
     use crate::operator_commands_dashboard::routes::*;
-    use crate::operator_commands_dashboard::routes::TmuxSession;
+    use crate::operator_commands_dashboard::tmux::TmuxSession;
     use crate::operator_commands_dashboard::distributed::remote_vms_from_hosts;
     use crate::operator_commands_dashboard::current_work::{format_recent_actions_for_cycle, read_recent_cycle_reports};
     use serde_json::json;
