+++ src/operator_commands_dashboard/tests_routes_b.rs
@@ -1,8 +1,9 @@
 #[cfg(test)]
 mod tests_b {
     use crate::operator_commands_dashboard::routes::*;
+    use crate::operator_commands_dashboard::tmux::parse_tmux_sessions;
     use crate::operator_commands_dashboard::memory::{build_agent_graph, classify_agent_layer};
-    use crate::operator_commands_dashboard::routes::TmuxSession;
+    use crate::operator_commands_dashboard::tmux::TmuxSession;
     use serde_json::json;
     #[test]
     fn sanitize_agent_name_rejects_invalid_names() {
